Stabilize sort of small record set
Summary: Filesort adds the record position in the table at the end of each key record in case addon fields are not added. However this position should not be used for sorting otherwise it adds an element of instability for equal keys (and it is not used to stabilize sort order - it would have to be added in big endian for that). The code in `Filesort_buffer::sort_buffer` decrements the key length to exclude the record position, and it uses decreased length in most cases except where it sorts a small record set using quicksort.

Reviewed By: hermanlee

Differential Revision: D25283386

fbshipit-source-id: 597da3abb24