Make rocksdb alter table inplace for changing default column value.
Summary:
We noticed a divergent behavior with respect to enums in innodb vs
rocksdb. This is because innodb treats alter table with compatible
enum as inplace. But rocksdb always does alter by copy.

With in-place alter, any insert does by RBR on the secondary instance
is retained. This means that even if secondary sees enums that it
does not yet recognize, it will still store that, until the alter runs on
the standby.

The fix makes the rocksdb behavior similar to innodb, by allowing
inplace alter for ALTER_COLUMN_DEFAULT. The alter statement for
compatible enums pass in ALTER_COLUMN_DEFAULT. Keeping
this under system variable rocksdb_alter_column_default_inplace

Reviewed By: yizhang82

Differential Revision: D25300341

fbshipit-source-id: 1339834e414