Add db id prefix in MyRocks key format
Summary:
The change adds db id prefix to MyRocks key. The
current MyRocks key format uses index number as the
lead prefix of the key. This index number ensures that all
the keys of a particular table/object are colocated in LSM.

However certain shard operations like (migration, backup)
operate on per database basis. In order to optimize these
operations for MyRocks, it becomes imperative to
colocate all the tables in database as well. This ensures
that entire data of a database is physically contiguous and
serves the foundation for efficient shard operations on
MyRocks.

Implementation notes:
1. Every database is assigned an id. This is created from
the hash of database name. it is a 4 byte unsigned int value.
2. A new enum INDEX_ID_SIZE is combination
of Index number size and db number size (4 bytes each).
Rdb_key_def caches db number for every table ref
in its struct,
3. The database name to dbid mapping is stored in data
dictionary of MyRocks. A new data dictionary entry
DB_ENTRY is added to persist the mapping.
4. Existing data dictionary keys have been modified to
store db_num for every index. This led to change in key
format of some of the key types. More info in comment
section of rdb_datadic.h
5. DDL_ENTRY_INDEX_VERSION is upgraded to new
version. This change is incompatible with previous
MyRocks database format. So upgrade will not be in-place.
Rdb_dict_manager::check_ddl_entries is added that acts
as a guard to prevent opening database in earlier format.
6. Had to refresh some tests to work with the new key
format.

Reviewed By: yizhang82

Differential Revision: D26387867

fbshipit-source-id: def5dc608a7