[mysql-test-run] Always reinitialize datadir when bootstrap options are present
Summary:
In 8.0, mtr was changed so that if bootstrap options exist and have not changed, then mysqld will not reinitialize the datadir. If bootstrap options do not exist, then the datadir
is cleaned by copying over an existing default installation with no bootstrap options.

Because we use the `--fast` mtr option to kill the server at the end of the test, there's no guarantee that a server has cleanly shutdown from the prior test case. This meant that if the prior test case did not have matching write policies and the two tests had the same bootstrap options, then the next test case would fail because we do not reinitialize the database.

The fix is to restore 5.6 behaviour by always reinitializing. There aren't a lot of tests with bootstrap options so the performance impact should be small. Also, this brings parity to tests without bootstrap options which will always get a clean datadir anyway.

Differential Revision: D25452410