Fix flakiness of replication_lag_auto_throttling test
Summary:
The test was flaky due to various sleeps to sync throttling cycles.
Instead of relying on sleeps I have added a debug variable to cycle through write throttling logic and much faster now. The test should be deterministic now.

Same as the fix for mysql 8.0 D25348677

In addition, I also fixed bug in counting write_throttling_logs for tagged queries. We were overcounting because the counter logging was done before deciding whether the query should be throttled or not.

Reviewed By: george-reynya

Differential Revision: D25468504

fbshipit-source-id: cfe2e0fdd60