Checking for inited bool to make sure global_init_info was successful
Summary:
A master.info and relay.info file can be present
but needs to be properly inited for use. We were bypassing the inited
check which could lead to issues in Raft.

In case there is an error in global_init_info, Raft will do a
raft_reset_slave and make another attempt at it. If both recourses
fail, the init of the plugin would fail.

Reviewed By: bhatvinay

Differential Revision: D25584004

fbshipit-source-id: 8811af60eff