Prevent underflow of m_n_mysql_tables_in_use
Summary:
The `m_n_mysql_tables_in_use` counter is underflowing during ADD/DROP partition. It underflows during ADD because `::close()` unlocks partitions that were never locked in the first place. The fix is to remove the unlock in `::close()`.

The counter underflows during DROP because dropped partitions are unlocked in `::commit_inplace_alter_table()` and a second time in `::external_lock()`. The fix is to remove the partitions from the `m_file` array so that in the second call to `::external_lock()`, only the remaining partitions are unlocked (instead unlocking all partitions, including ones that were already dropped).

The consequence of this bug is that some statements may commit prematurely (once count drops to 0) even though the statement may hit errors later before the final `external_lock` call is made.

Squash with: D23177984 D23266554

Reviewed By: luqun

Differential Revision: D25592121

fbshipit-source-id: c4f7f266267