Fix SSL memory leak on connection close
Summary:
This bug affects the async client case, in the event of a connection timeout.

In some cases, we could be initiating a VIO shutdown and cleanup on a
VIO struct that has a non-null ssl_arg using the fallback vio_delete()
function instead of the vio_ssl_delete() function. This causes a memory
leak on the SSL struct because nothing is freeing it.

The ssl_arg is deleted in the call to vio_ssl_delete(), which gets called if the
type == VIO_TYPE_SSL. However in some cases, the type we get is VIO_TYPE_TCPIP
(probably some transient state, interrupted from becoming VIO_TYPE_SSL
maybe by the timeout), and because of that, we call vio_delete() instead of vio_ssl_delete(),
which doesn't free up the SSL data.

Upstream bug: https://bugs.mysql.com/bug.php?id=102189

Reviewed By: lloyd

Differential Revision: D25814364

fbshipit-source-id: 7f72d3100e2