Add partial index iterator
Summary:
This adds the partial index iterator. It is a special iterator that sorts groups from the primary key on the fly as needed, and if it exceeds a certain threshold, it will materialize the rows on the secondary key as well.

In order to protect against concurrent index updates and materialization, locks are taken on the group prefix on the secondary key. When updating an partial index, a shared lock on the group is held, and a read is performed to see if a SK write is necessary or not. The shared lock is held until transaction commit. When materializing a group, an exclusive lock on the group is held, in order to ensure that we don't miss any writes going to that group.

For point queries, the secondary key read is not found, we simply extract the primary key parts and read from the primary key. This means that point queries don't trigger materialization though.

Reviewed By: hermanlee

Differential Revision: D25933178