Fix bypass range query
Summary:
In current bypass implementation, its range query implementation is not correct. For A >= begin and A <= end , it'll only pick begin *or* end as start of the range, but it will only use the prefix as the end of the range, which means it can potentially scan a lot more rows than needed. It rely on the condition A >= begin / A <= end to filter out the unnecessary rows so the end result is correct (that's why this issue is not discovered in testing).

The correct way to implement is to set up (begin, end) range slice correctly, and do NOT rely on any conditional evaluation to filter rows. The tricky part is to determine the correct begin/end based on a few factors:
* forward / reverse column family
* ascending / descending orders
* inclusive vs non-inclusive ranges (>=, <= vs <, >)
* prefix query vs range query

For better testing, I've done the following:
* verify_bypass_query.inc that will run the query in bypass and non-bypass, verify the row reads are same or less than non-bypass case to ensure such issue won't happen again. Most bypass query validation are moved to use verify_bypass_query.inc. We can also add more validation in the future in verify_bypass_query.inc as needed. As a bonus, it'll make verifying the results are the same in bypass/non-bypass case easy as well.
* move range key validation from bypass_select_scenarios into bypass_select_range_pk and bypass_select_range_sk
* added more test cases to bypass_select_range_pk / bypass_select_range_sk, especially for PK case where some scenarios are missing
* dump query results to file and verify query results didn't change w/ and w/o bypass. I had to back port `output` mysqltest support to make this work properly (for some reason --exec $MYSQL doesn't like multiline queries".

For review, there is no need to go through results file as there are validation in-place to make sure query return same results w/ and w/o bypass.

Reviewed By: luqun

Differential Revision: D25982322

fbshipit-source-id: c857f9a382e