Fix main.all_persisted_variables
Summary:
Our asan builds do not have jemalloc compiled in and this is causing the `main.all_persisted_variables` test to fail.

It seems like there was supposed to be some accounting in the test to adjust for this, but it's not clear to me how this ever worked. It seems like we should only be adjusting the number down, because if an extra variable exists, the `SELECT COUNT(*)` statement that populates `$total_global_vars` should have already counted any extra variables.

Squash with: D19777409 (https://github.com/facebook/mysql-5.6/commit/df413c536a957ff58af49d8dba0dd86ea502437c)

Reviewed By: yizhang82

Differential Revision: D26024689

fbshipit-source-id: 4440907057b