Fix SSL memory leak on connection close in a better way
Summary:
This fixes an issue with a previous change (D25814364). The problem that the previous change tried to address was a leak during timeout in the async path.

When timeout occurs in the async path, the caller does not re-enter the nonblocking functions to perform cleanup and rely on destructors to do the right thing. In this case, we were not freeing the SSL object that was allocated, and so we modified `internal_vio_delete` to do the deletion.

However, this causes problems for the sync path because the sync path calls `SSL_free(vio->ssl_arg)` if `ssl_handshake_loop` returns an error, and then on vio destruction, we call `SSL_free` again on the same SSL object.

Instead, we should be freeing `ctx->ssl` instead of `vio->ssl_arg` to match the behaviour of the sync path. The `ctx->ssl` variable is supposed to mimic the SSL stack variable, and hold the SSL object until SSL is complete and is transferred to the vio object.

Squash with: D25814364

Reviewed By: fadimounir, aditya-jalan

Differential Revision: D26028251

fbshipit-source-id: 18f092ed87c