Properly fallback if executing unsupported case
Summary: In most cases bypass parser `select_parser` should identify unsupported case and fallback to regular query. However, there are a few checks that are expensive and is done as part of execution, and we should properly fallback in those case as well. In a future version maybe we should refactor it so that those are another phase of parsing as well, which would also help bypass RPC.

Reviewed By: Pushapgl

Differential Revision: D26033752

fbshipit-source-id: 54e1ce844bf