rocksdb system variables to change db number/index number
Summary:
The change implements 2 system variables rocksdb_alter_current_db_num
and rocksdb_alter_index_num that change the db number, index
numbers of the table respectively.
These system variables are dumped by mysqldump.
rocksdb_dump_internal_id is provided as a optional parameter to
myslqdump.

rocksdb_alter_current_db_num is a pretty simple implementation that
changes the db number of the target database to be the same as
the source on which the mysqldump ran.

rocksdb_alter_index_num is passsed as a string in form of json
```
{
  "table":<table_name>,
  "autoinc":<int value> ,
  "indexes":[{"idx_type":<type>, "idx_name":<index name>, "idx_num":<index number>} ... ]
}
```
mysqldump dumps all the index numbers of the table in above format.
This can then be executed on the target database.

The above changes ensure that target database internal myrocks
schema is same as source database. This opens up the possibility
of efficient database level operations (migration, backup) in myrocks.

Reviewed By: yizhang82

Differential Revision: D27217979

fbshipit-source-id: 86364c42faf