cur_log_ext should be incremented only on reaching consensus
Summary:
porting D24442402

The plugin and the server share the cur_log_ext variable
pointer. Plugin uses this variable to create the dense index entries
for each message. So the cur_log_ext should only be updated, when we
are sure that append entries requests will go to the next file and not
to the current file. This can only be done safely when consensus is
reached on the rotate event. Otherwise the before_commit hook could
have failed, then we received a truncation request on same file, and
then we could be forced to append another No-Op to the same file, i.e.
safety comes only after consensus commit.

Reviewed By: Pushapgl

Differential Revision: D26156483

fbshipit-source-id: 1ba44dbc106