Gradual throttling
Summary:
This diff adds a gradual throttling logic to the existing replication lag auto throttling system.

Note that, It does not affect manual throttling system. Any manual throttling rule will throttle 100% of the requests as soon as it is added.

Added a new server variable `write-throttle-rate-step` which determines the step by which throttle rate is stepped up or down. It must be a number between 0-100. It defaults to 100 which means throttle rate is stepped up from 0% to 100% or it is stepped down from 100% to 0% which is the default behavior now.

Added a new column to `I_S.write_throttling_rules` to display `throttle_rate`

In addition, I also fixed a bug with clearing `auto_throttling_rules`, `currently_throttled_entities` & `currently_monitored_entity` when throttling rules are cleared or auto throttling is turned off. Previously, we were not clearing these data structures properly which would have caused some inconsistencies in the algorithm.

Reviewed By: george-reynya

Differential Revision: D26217123

fbshipit-source-id: 9724ec5ae94