Support backwards compatible compression on replication
Summary:
8.0.20 has support for zstd on the replication channel, but
FB's 8.0.17 and 5.6.35 uses the compression_lib connection attribute to
indicate which compression type to use.

Add backwards compatibility so that if the attribute is present, then
it will be used to determine compression type. Once everything is
upgraded to 8.0.20, then this patch can be dropped.

8.0.20 also supports specifying the zstd compression level from the
client. However, the 8.0.17 implementation used the compression
level on the server, so client requested compression level is ignored if
the compression_lib attribute is specified (i.e. this patch will enforce
using all compression level settings from the server).

While 8.0.20 allows selecting individual compression types through the
client flags and enabling/disabling server support for some compression
types, for backwards compatibility, this has been reduced to either
supporting all types, or support none of the compression types.

Certains tests needed to be rewritten and re-recorded for this change.

The main changes are as follows:

1. Modify server handshake to include CLIENT_COMPRESS if any compression
   type is supported.
2. Modify client handshake reply to include the compression_lib attribute
   to indicate the type of compression supported if compression is
   requested. The CLIENT_COMPRESS bit is only set if the server does
   not advertise any of the new compression settings.
3. Modify the server parsing of the handshake to use the compression_lib
   attribute to designate the compression type. If the compression_lib
   attribute is not present, then fall back to normal compression
   negotiation.

Reviewed By: lth

Differential Revision: D26472664

fbshipit-source-id: 53d33318f77