Fix mysql range query optimizer bug in key_or
Summary:
For range query case with multiple OR, optimizer rely on `tree_or` and `key_or` to 'merge' the OR sub-tree together.

And in the case of having multiple ranges that are the same, `key_or` would see that both ranges key1 and key2 are exactly the same, release key2->next_key_part, and then OR key1->next_key_part and key2->next_key_part= NULL (since it is released) with `key_or`. However, because NULL sub-tree are treat as TRUE (same as is_always), the entirety of key2->next_key_part gets dropped, so you end up with just the first key.

It makes sense for NULL sub-tree to be treat as TRUE for `key_or`, because for example (A > 1 AND NULL) OR (A > 1 AND B > 1 AND NULL) should be merged as (A > 1 AND NULL), so it makes sense for NULL to be treated as TRUE so that `key_or(NULL, key)` should come back as NULL. So the right way to fix this is to stop handling current range and simply move to the next range (without falling through to the remaining logic).

Because of this bug the range query plan end up using much less keys and end up being more expensive, so usually a ref plan gets picked with less keys, and the query becomes much more expensive.

Upstream bug: https://bugs.mysql.com/bug.php?id=102634

Reviewed By: lth, Pushapgl

Differential Revision: D26477017

fbshipit-source-id: 44536b611a5