fix in metadata event type read funtion
Summary:
Fix to avoid signed extesnion when the value_length is a large number (such
that the MSB is set). This happens in a config change event when the
length of the config change string is large (causing the MSB in the
16bit unsigned integer to be set). While reading this back we have to
ensure the signed conversion is done correctly.

For example, a config change string of length 45460 bytes (0xb194) gets
sign extended to 65428 bytes (0xff94)

Reviewed By: anirbanr-fb

Differential Revision: D26496395

fbshipit-source-id: 32055e24e03