Avoid passing nullptr to memcmp in Rdb_string_writer comparisons
Summary: For empty Rdb_string_writer, m_data is an empty std::vector<uchar> and data() returns nullptr which ends up being passed to memcmp. In either case we do pass length=0 to memcmp so that's most likely why this still works in release, but it did trigger an ubsan warning in 8.0 (which I'm fixing together with the port) but didn't in 5.6 for some reason. This fixes it in 5.6

Reviewed By: luqun

Differential Revision: D26572454

fbshipit-source-id: 1a70b717b4f