Change float comparison semantics for BI consistency check
Summary:
It is a known problem that a mysqldump/restore from 5.6 to
8.0 instance causes a loss of precision of floating point
values. This is because mysqldump dumps float only until
6 decimal places. But internally the float stores higher
precision values. Returning 6 decimal precision for float is
the behavior of MySQL by design and it cannot be changed
unless we specify precision in float definition.
This loss of precision causes BI image consistency check
to fail when 5.6 is primary and newly created 8.0 instance
is secondary replicating from 5.6 instance. The BI image in
binlog contains the high precision floating values stored in
5.6. But 8.0 stores the less precise values created by
mysqldump.
This can also happen vice versa where 5.6 is secondary
replicating from 8.0 primary and is created by mysqldump/restore.
The fix is to change the BI comparison logic for float so
that we only compare the first 6 decimal places of float
datatypes.
Squash with D25958618 (https://github.com/facebook/mysql-5.6/commit/66fdd81edb59e6de7962ee9184c9e5b248e6381c)

Reviewed By: hermanlee

Differential Revision: D26657722

fbshipit-source-id: 9dbacdb8a59