Changes required for MySQL privacy plugin
Summary:
MySQL plugin for privacy (D26704751) is build using the AUDIT plugin. The plugin requires access to the from tables of the query. This is extracted from the LEX tree produced by parser after the parse step is complete.

This diff adds a function to sql/sql_thd_internal_api.h/.cc:

- thd_get_query_tables() - Returns the list of tables extracted from the LEX tree produced by parser.

Reviewed By: lth, george-reynya

Differential Revision: D26704727

fbshipit-source-id: 4bb8d7dc975