BUG#32655364: MEMORY LEAK IN SOURCE_IO_MONITOR
Summary:
Source_IO_monitor has a static shared instance that was not being
properly terminated. The root issue was the instance being static,
suffering from out of order initialization and destruction, in
particular the external performance schema instrumentation global
references it pointed to were cleaned before it.

To solve the above issue, now this instance is a singleton on
mysqld.cc with clear defined initialization and destruction points.

Porting notes: Backport memory leak fix of Source_IO_monitor for valgrind runs.
Drop when porting forward to 8.0.27

Squash with null

Reviewed By: Pushapgl

Differential Revision: D33359868

fbshipit-source-id: 843ca5ed428