handle partial rotate events correctly on rstart
Summary:
Raft adds a metadata event before a rotate event in the binlog. The
metadata event contains the (term, index) corresponding to the rotate
'operation' [rotate is considered a raft operation and it goes through
consensus]. On restarts, any partial events written (only metadata
event, but no rotate event) towards the end of the file needs to be
handled correctly.

Reviewed By: anirbanr-fb

Differential Revision: D27063119

fbshipit-source-id: 5a3fcc80cf7