Removing the mutex from the signal side of ABS
Summary:
the mutex LOCK_lack_acked can be heavily contended.
This can be especially the case when there are lots of async tailers.
In this diff, we use the fact that the last acked file num and pos pair
is an atomic and therefore does not need the additional mutex for
writes.

Since the conditional variable is a timed wait in a loop, the missed
signal problem is not an issue.

Reviewed By: abhinav04sharma

Differential Revision: D27507281

fbshipit-source-id: c4fa604aab7